+++ daemon/containerd/image_delete.go
@@ -1,3 +1,6 @@
+// FIXME(thaJeztah): remove once we are a module; the go:build directive prevents go from downgrading language version to go1.16:
+//go:build go1.23
+
 package containerd
 
 import (
@@ -9,6 +12,7 @@ import (
 	c8dimages "github.com/containerd/containerd/v2/core/images"
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
+	"github.com/containerd/platforms"
 	"github.com/distribution/reference"
 	"github.com/docker/docker/api/types/events"
 	imagetypes "github.com/docker/docker/api/types/image"
@@ -17,6 +21,7 @@ import (
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/internal/metrics"
 	"github.com/docker/docker/pkg/stringid"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
 // ImageDelete deletes the image referenced by the given imageRef from this
@@ -26,11 +31,11 @@ import (
 // imageRef is a repository reference or not.
 //
 // If the given imageRef is a repository reference then that repository
-// reference will be removed. However, if there exists any containers which
+// reference is removed. However, if there exists any containers which
 // were created using the same image reference then the repository reference
 // cannot be removed unless either there are other repository references to the
 // same image or force is true. Following removal of the repository reference,
-// the referenced image itself will attempt to be deleted as described below
+// the referenced image itself is attempted to be deleted as described below
 // but quietly, meaning any image delete conflicts will cause the image to not
 // be deleted and the conflict will not be reported.
 //
@@ -47,18 +52,25 @@ import (
 // The image cannot be removed if there are any hard conflicts and can be
 // removed if there are soft conflicts only if force is true.
 //
-// If prune is true, ancestor images will each attempt to be deleted quietly,
+// If prune is true, ancestor images are attempted to be deleted quietly,
 // meaning any delete conflicts will cause the image to not be deleted and the
 // conflict will not be reported.
 //
 // TODO(thaJeztah): image delete should send prometheus counters; see https://github.com/moby/moby/issues/45268
-func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, prune bool) (response []imagetypes.DeleteResponse, retErr error) {
+func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, options imagetypes.RemoveOptions) (response []imagetypes.DeleteResponse, retErr error) {
 	start := time.Now()
 	defer func() {
 		if retErr == nil {
 			metrics.ImageActions.WithValues("delete").UpdateSince(start)
 		}
 	}()
+
+	if len(options.Platforms) > 0 && !options.Force {
+		return nil, cerrdefs.ErrInvalidArgument.WithMessage("Content will be removed from all images referencing this variant. Use —-force to force delete.")
+	}
+
+	force := options.Force
+	prune := options.PruneChildren
 
 	var c conflictType
 	if !force {
@@ -96,25 +108,18 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 			c &= ^conflictActiveReference
 		}
 		if named != nil && len(sameRef) > 0 && len(sameRef) != len(all) {
-			var records []imagetypes.DeleteResponse
-			for _, ref := range sameRef {
-				// TODO: Add with target
-				err := i.images.Delete(ctx, ref.Name)
-				if err != nil {
-					return nil, err
-				}
-				if nn, err := reference.ParseNormalizedNamed(ref.Name); err == nil {
-					familiarRef := reference.FamiliarString(nn)
-					i.logImageEvent(ref, familiarRef, events.ActionUnTag)
-					records = append(records, imagetypes.DeleteResponse{Untagged: familiarRef})
-				}
+			if len(options.Platforms) > 0 {
+				return i.deleteImagePlatforms(ctx, img, imgID, options.Platforms)
 			}
-			return records, nil
+			return i.untagReferences(ctx, sameRef)
 		}
 	} else {
 		imgID = image.ID(img.Target.Digest)
 		explicitDanglingRef := strings.HasPrefix(imageRef, imageNameDanglingPrefix) && isDanglingImage(*img)
 		if isImageIDPrefix(imgID.String(), imageRef) || explicitDanglingRef {
+			if len(options.Platforms) > 0 {
+				return i.deleteImagePlatforms(ctx, img, imgID, options.Platforms)
+			}
 			return i.deleteAll(ctx, imgID, all, c, prune)
 		}
 		parsedRef, err := reference.ParseNormalizedNamed(img.Name)
@@ -127,20 +132,10 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 			return nil, err
 		}
 		if len(sameRef) != len(all) {
-			var records []imagetypes.DeleteResponse
-			for _, ref := range sameRef {
-				// TODO: Add with target
-				err := i.images.Delete(ctx, ref.Name)
-				if err != nil {
-					return nil, err
-				}
-				if nn, err := reference.ParseNormalizedNamed(ref.Name); err == nil {
-					familiarRef := reference.FamiliarString(nn)
-					i.logImageEvent(ref, familiarRef, events.ActionUnTag)
-					records = append(records, imagetypes.DeleteResponse{Untagged: familiarRef})
-				}
+			if len(options.Platforms) > 0 {
+				return i.deleteImagePlatforms(ctx, img, imgID, options.Platforms)
 			}
-			return records, nil
+			return i.untagReferences(ctx, sameRef)
 		} else if len(all) > 1 && !force {
 			// Since only a single used reference, remove all active
 			// TODO: Consider keeping the conflict and changing active
@@ -150,7 +145,17 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 
 		using := func(c *container.Container) bool {
 			if c.ImageID == imgID {
-				return true
+				if len(options.Platforms) == 0 {
+					return true
+				}
+				for _, p := range options.Platforms {
+					pm := platforms.OnlyStrict(p)
+					if pm.Match(c.ImagePlatform) {
+						return true
+					}
+				}
+
+				// No match for the image reference, but continue to check if used as mounted image
 			}
 
 			for _, mp := range c.MountPoints {
@@ -182,6 +187,10 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 				return nil, err
 			}
 
+			if len(options.Platforms) > 0 {
+				return i.deleteImagePlatforms(ctx, img, imgID, options.Platforms)
+			}
+
 			// Delete all images
 			err := i.softImageDelete(ctx, *img, all)
 			if err != nil {
@@ -194,7 +203,69 @@ func (i *ImageService) ImageDelete(ctx context.Context, imageRef string, force, 
 		}
 	}
 
+	if len(options.Platforms) > 0 {
+		return i.deleteImagePlatforms(ctx, img, imgID, options.Platforms)
+	}
 	return i.deleteAll(ctx, imgID, all, c, prune)
+}
+
+// deleteImagePlatforms iterates over a slice of platforms and deletes each one for the given image.
+func (i *ImageService) deleteImagePlatforms(ctx context.Context, img *c8dimages.Image, imgID image.ID, platformsToDel []ocispec.Platform) ([]imagetypes.DeleteResponse, error) {
+	var accumulatedResponses []imagetypes.DeleteResponse
+	for _, p := range platformsToDel {
+		responses, err := i.deleteImagePlatformByImageID(ctx, img, imgID, &p)
+		if err != nil {
+			return nil, fmt.Errorf("failed to delete platform %s for image %s: %w", platforms.Format(p), imgID.String(), err)
+		}
+		accumulatedResponses = append(accumulatedResponses, responses...)
+	}
+	return accumulatedResponses, nil
+}
+
+func (i *ImageService) deleteImagePlatformByImageID(ctx context.Context, img *c8dimages.Image, imgID image.ID, platform *ocispec.Platform) ([]imagetypes.DeleteResponse, error) {
+	pm := platforms.OnlyStrict(*platform)
+	var target ocispec.Descriptor
+	if img == nil {
+		// Find any image with the same target
+		// We're deleting by digest anyway so it doesn't matter - we just
+		// need a c8d image object to pass to getBestPresentImageManifest
+		i, err := i.resolveImage(ctx, imgID.String())
+		if err != nil {
+			return nil, err
+		}
+		img = &i
+	}
+	imgMfst, err := i.getBestPresentImageManifest(ctx, *img, pm)
+	if err != nil {
+		return nil, err
+	}
+	target = imgMfst.Target()
+
+	var toDelete []ocispec.Descriptor
+	err = i.walkPresentChildren(ctx, target, func(ctx context.Context, d ocispec.Descriptor) error {
+		toDelete = append(toDelete, d)
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// TODO: Check if these are not used by other images with different
+	// target root images.
+	// The same manifest can be referenced by different image indexes.
+	var response []imagetypes.DeleteResponse
+	for _, d := range toDelete {
+		if err := i.content.Delete(ctx, d.Digest); err != nil {
+			if cerrdefs.IsNotFound(err) {
+				continue
+			}
+			return nil, err
+		}
+		if c8dimages.IsIndexType(d.MediaType) || c8dimages.IsManifestType(d.MediaType) {
+			response = append(response, imagetypes.DeleteResponse{Deleted: d.Digest.String()})
+		}
+	}
+	return response, nil
 }
 
 // deleteAll deletes the image from the daemon, and if prune is true,
@@ -405,6 +476,24 @@ func (idc *imageDeleteConflict) Error() string {
 }
 
 func (*imageDeleteConflict) Conflict() {}
+
+// untagReferences deletes the given image references and returns the appropriate response records
+func (i *ImageService) untagReferences(ctx context.Context, refs []c8dimages.Image) ([]imagetypes.DeleteResponse, error) {
+	var records []imagetypes.DeleteResponse
+	for _, ref := range refs {
+		// TODO: Add with target
+		err := i.images.Delete(ctx, ref.Name)
+		if err != nil {
+			return nil, err
+		}
+		if nn, err := reference.ParseNormalizedNamed(ref.Name); err == nil {
+			familiarRef := reference.FamiliarString(nn)
+			i.logImageEvent(ref, familiarRef, events.ActionUnTag)
+			records = append(records, imagetypes.DeleteResponse{Untagged: familiarRef})
+		}
+	}
+	return records, nil
+}
 
 // checkImageDeleteConflict returns a conflict representing
 // any issue preventing deletion of the given image ID, and