+++ integration/image/remove_test.go
@@ -4,10 +4,14 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/containerd/platforms"
+
 	containertypes "github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/integration/internal/container"
+	"github.com/docker/docker/internal/testutils/specialimage"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
 	"gotest.tools/v3/skip"
@@ -90,4 +94,84 @@ func TestRemoveByDigest(t *testing.T) {
 	inspect, err = client.ImageInspect(ctx, "test-remove-by-digest")
 	assert.Check(t, is.ErrorType(err, errdefs.IsNotFound))
 	assert.Check(t, is.DeepEqual(inspect, image.InspectResponse{}))
+}
+
+func TestRemoveWithPlatform(t *testing.T) {
+	skip.If(t, !testEnv.UsingSnapshotter())
+
+	ctx := setupTest(t)
+	apiClient := testEnv.APIClient()
+
+	imgName := strings.ToLower(t.Name()) + ":latest"
+
+	platformHost := platforms.Normalize(ocispec.Platform{
+		Architecture: testEnv.DaemonInfo.Architecture,
+		OS:           testEnv.DaemonInfo.OSType,
+	})
+	someOtherPlatform := platforms.Platform{
+		OS:           "other",
+		Architecture: "some",
+	}
+
+	var imageIdx *ocispec.Index
+	var descs []ocispec.Descriptor
+	specialimage.Load(ctx, t, apiClient, func(dir string) (*ocispec.Index, error) {
+		idx, d, err := specialimage.MultiPlatform(dir, imgName, []ocispec.Platform{
+			platformHost,
+			{
+				OS:           "linux",
+				Architecture: "test", Variant: "1",
+			},
+			{
+				OS:           "linux",
+				Architecture: "test", Variant: "2",
+			},
+			someOtherPlatform,
+		})
+		descs = d
+		imageIdx = idx
+		return idx, err
+	})
+	_ = imageIdx
+
+	for _, tc := range []struct {
+		platform *ocispec.Platform
+		deleted  ocispec.Descriptor
+	}{
+		{&platformHost, descs[0]},
+		{&someOtherPlatform, descs[3]},
+	} {
+		resp, err := apiClient.ImageRemove(ctx, imgName, image.RemoveOptions{
+			Platforms: []ocispec.Platform{*tc.platform},
+			Force:     true,
+		})
+		assert.NilError(t, err)
+		assert.Check(t, is.Len(resp, 1))
+		for _, r := range resp {
+			assert.Check(t, is.Equal(r.Untagged, ""), "No image should be untagged")
+		}
+		checkPlatformDeleted(t, imageIdx, resp, tc.deleted)
+	}
+
+	// Delete the rest
+	resp, err := apiClient.ImageRemove(ctx, imgName, image.RemoveOptions{})
+	assert.NilError(t, err)
+
+	assert.Check(t, is.Len(resp, 2))
+	assert.Check(t, is.Equal(resp[0].Untagged, imgName))
+	assert.Check(t, is.Equal(resp[1].Deleted, imageIdx.Manifests[0].Digest.String()))
+	// TODO: Should it also include platform-specific manifests?
+}
+
+func checkPlatformDeleted(t *testing.T, imageIdx *ocispec.Index, resp []image.DeleteResponse, mfstDesc ocispec.Descriptor) {
+	for _, r := range resp {
+		if r.Deleted != "" {
+			if assert.Check(t, is.Equal(r.Deleted, mfstDesc.Digest.String())) {
+				continue
+			}
+			if r.Deleted == imageIdx.Manifests[0].Digest.String() {
+				t.Log("Root image was deleted, expected only platform:", platforms.FormatAll(*mfstDesc.Platform))
+			}
+		}
+	}
 }